# Improves FLANN's heap allocations by a memory pool