FLANN allocates huge amounts of memory with every call to `findNeighbors` due to its private heap allocations. This affects memory throughput and overall speed of user applications, especially the ones that frequently query for neighbors. After all, although `malloc` does cache on its own, it also does a lot of smart things besides that, which costs some time penalty. To solve the problem, this PR introduces a static, thread-safe, and bounded memory pool for heap allocations. It have been tested thoroughly for very long hours, under extensive parallelization, and with strict leakage profiling. **The tests showed more than 70x memory throughput gain**. For roughly the same amount of calls, the memory footprint of `findNeighbors` using master branch was around 18.2 GB/sec, which got reduced by this PR to just 250.5 MB/sec (see the allocations throughput table below). Additionally, this PR cleans up some unnecessary variables in the `cvflann::Heap` class.

| Master | This PR |
|----------|---------|
| ![master](https://user-images.githubusercontent.com/42140441/126982801-304ac0b6-e8a9-48a6-a8c6-369f0fdc5a45.png) | ![PR](https://user-images.githubusercontent.com/42140441/126982824-b4ff41f2-9e18-4925-a7ef-95901efe785d.png) |

### Pull Request Readiness Checklist

<!-- See details at https://github.com/opencv/opencv/wiki/How_to_contribute#making-a-good-pull-request -->

- [x] I agree to contribute to the project under Apache 2 License.
- [x] To the best of my knowledge, the proposed patch is not based on a code under GPL or other license that is incompatible with OpenCV
- [x] The PR is proposed to proper branch
- N/A: There is reference to original bug report and related work

#### Already existing:
- [x] There is accuracy test, performance test and test data in opencv_extra repository, if applicable
      Patch to opencv_extra has the same branch name.
- [x] The feature is well documented and sample code can be built with the project CMake


```
force_builders=linux32
```