@cv3d 2021-08-05
@alalek I revised according to your valuable comments.

 Please let me know if you still think there are some remaining improvement that can be made.
Thanks~