@cv3d 2021-08-09
@alalek Appreciate your followup.
Maybe you want to go through the resolutions I made, and double check they are satisfactory. Thanks~