@alalek 2021-08-10
Please take a look on warnings from 32-bit compilation on Linux platforms.